move download_url, add benchmarking
