+++ environment.yml
@@ -34,3 +34,4 @@ dependencies:
         - sphinx-rtd-theme==1.0.0
         - sphinx-autoapi==1.8.4
         - sphinx-copybutton==0.4.0
+        - tqdm
