+++ setup.py
@@ -35,6 +35,7 @@ setuptools.setup(
         "torch",
         "opencv-contrib-python",
         "python-bioformats",
+        "tqdm",
     ],
     classifiers=[
         "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
